Zhangchar/refactor input component (#26)
* create user input element

* added missing items in component.json

* add data input element

* add number input element, passing mrp type to i18n

* add text input element

* add percentage input element

* removing warnings

* update UNSAFE_ lifecycle functions

* update UNSAFE_ lifecycle functions

* added placeholder color for textarea, add procurment type input
