+++ src/hello-world/public/locales/zh-CN/component.json
@@ -30,9 +30,9 @@
     "lead-time": "部件交货时间",
     "description": "描述",
     "note": "备注",
-    "team-name": "* 产生班组名",
+    "team-name": "产生班组名",
     "shift":"* 工作班次",
-    "updated-by-user": "* 记录人姓名:",
+    "updated-by-user": "输入姓名:",
     "given-user-name": "* 名",
     "family-user-name": "* 姓",
     "actual-quantity-per-shift": "* 该工作班次实际产生/使用的件数",
