+++ src/hello-world/src/stylesheets/ccilab/scss/components/_ccilab-component-list.scss
@@ -25,9 +25,23 @@
  color:$cci-hint-red;
 }
 
-::-ms-input-placeholder { /* Microsoft Edge */
+::-webkit-input-placeholder { /* Microsoft Edge */
  color:$cci-hint-red;
 }
+
+textarea::placeholder {
+  color:$cci-hint-red;
+  opacity: 1; /* Firefox */
+}
+
+textarea:-ms-input-placeholder { /* Internet Explorer 10-11 */
+  color:$cci-hint-red;
+ }
+ 
+ textarea::-webkit-input-placeholder { /* Microsoft Edge */
+  color:$cci-hint-red;
+ }
+
 //// controlled using inline style 
 //  .flyout-menu_hide {
 //    // -webkit-transform: translate3d(-10vw, 0, 0);
